nextauth refactoring
