+++ next-auth.d.ts
@@ -5,6 +5,7 @@ import 'next-auth';
 declare module 'next-auth' {
   interface User extends Partial<LoginResponseTypeExtended> {
     jellyfinToken: string;
+    SessionInfo?: LoginResponseTypeExtended['SessionInfo'];
   }
 
   interface Session {
